new modular structure

+++ src/index.js
@@ -4,11 +4,11 @@ const Cron = require('cron').CronJob;
 
 const config = JSON.parse(fs.readFileSync(`${__dirname}/config.json`, "utf-8").toString());
 
-const client = new Discord.Client({ partials: ['USER'] });
+const client = new Discord.Client({ partials: ['USER', 'CHANNEL'] });
 
 const channelNames = ['🟢 Online: ', "👪 total Member: ", "🕒 mm:hh 📆 01.01.00"];
 
-let settings = JSON.parse(fs.readFileSync(`${__dirname}/settings.json`, "utf-8"));
+let managedGuilds = JSON.parse(fs.readFileSync(`${__dirname}/settings.json`, "utf-8"));
 
 
 let job = new Cron('0 * * * * *', async () => {
@@ -16,24 +16,37 @@ let job = new Cron('0 * * * * *', async () => {
         time = `${convertTime(d.getHours())}:${convertTime(d.getMinutes())}`,
         date = `${d.getDate()}/${d.getMonth() +1}/${d.getFullYear()}`;
 
-    for (e in settings){
+    managedGuilds.forEach(async e => {
         let guild = await client.guilds.fetch(e),
-            channel = guild.channels.cache.get(settings[e][2]);
+            channels = guild.channels.cache.filter(channel => channel.type == "voice"),
+            settings = {
+                type: "voice",
+                permissionOverwrites: [{
+                    id: guild.roles.everyone,
+                    deny: ['CONNECT']
+                }]
+            };
+            
+        channels.forEach(channel => {
+            let channelName = channel.name.split(' ');
+            try{
+                let time = channelName[1].split(':'),
+                    date = channelName[3].split('/');
 
-        channel.delete();
-        let newChannel = await guild.channels.create(`🕒 ${time} 📆 ${date}`, {
-            type: "voice",
-            parent: settings[e][3],
-            permissionOverwrites: [{
-                id: guild.roles.everyone,
-                deny: ['CONNECT']
-            }]
+                let result = parseFloat(time[0] + time[1] + date[0] + date [1] + date[2]);
+
+                if(result){
+                    if(channel.parentID) settings.parent = channel.parentID;
+
+                    settings.position = channel.position;
+
+                    channel.delete();
+                }
+            }catch(err){}
         });
 
-        settings[e][2] = newChannel.id;
-
-        fs.writeFileSync(`${__dirname}/settings.json`, JSON.stringify(settings));
-    }
+        guild.channels.create(`🕒 ${time} 📆 ${date}`, settings);
+    });
 });
 job.start();
 
@@ -53,14 +66,28 @@ client.on('ready', ()=> {
 });
 
 
-client.on('message', msg => {
+client.on('message', async msg => {
     if(!msg.content.startsWith(config.prefix)) return;
     msg.content = msg.content.slice(config.prefix.length).split(' ');
 
     if(msg.content[0] == "ping") msg.channel.send("pong");
 
     if(msg.content[0] == "test"){
-        msg.channel.send("no tests running!");
+    }
+
+    if(msg.content[0] == "cleanup"){
+        managedGuilds.forEach(async guildID => {
+            let guild = await client.guilds.fetch(guildID),
+                channels;
+
+            await channelNames.forEach(name => {
+                channels = guild.channels.cache.filter(channel => channel.name.startsWith(name));
+
+                channels.each(channel => {
+                    channel.delete();
+                });
+            });
+        });
     }
 
 
@@ -74,7 +101,8 @@ client.on('message', msg => {
         let guild = msg.guild,
             count = 3;
 
-        settings[guild.id] = [];
+        managedGuilds.push(guild.id);
+        fs.writeFileSync(`${__dirname}/settings.json`, managedGuilds);
 
         msg.guild.channels.create('stats', {type: "category"}).then(async category => {
             channelNames.forEach(async item => {
@@ -88,12 +116,8 @@ client.on('message', msg => {
                     }]
                 });
                 count--
-                settings[guild.id].push(channel.id);
 
                 if(count <= 0){
-                    settings[guild.id].push(category.id);
-                    fs.writeFileSync(`${__dirname}/settings.json`, JSON.stringify(settings));
-
                     checkOnlineCount(guild);
                     checkMemberCount(guild);
                 }
@@ -125,41 +149,44 @@ client.on('presenceUpdate', (oldUser, newUser) => {
 
 
 async function checkMemberCount(guild){
-    let oldChannel = await client.channels.fetch(settings[guild.id][1]);
-    oldChannel.delete();
+    let oldChannel = await guild.channels.cache.filter(channel => channel.name.startsWith(channelNames[1])),
+        settings = {
+            type: "voice",
+            permissionOverwrites: [{
+                id: guild.roles.everyone,
+                deny: ['CONNECT']
+            }]
+        };
 
-    let channel = await guild.channels.create(`${channelNames[1]} ${guild.memberCount}`, {
-        type: "voice",
-        parent: settings[guild.id][3],
-        position: 1,
-        permissionOverwrites: [{
-            id: guild.roles.everyone,
-            deny: ['CONNECT']
-            
-        }]
+    oldChannel.each(channel => {
+        if(channel.parentID) settings.parent = channel.parentID;
+
+        settings.position = channel.position;
+
+        channel.delete();
     });
 
-    settings[guild.id][1] = channel.id;
-
-    fs.writeFileSync(`${__dirname}/settings.json`, JSON.stringify(settings));
+    let channel = await guild.channels.create(`${channelNames[1]} ${guild.memberCount}`, settings);
 }
 async function checkOnlineCount(guild){
-    let oldChannel = await client.channels.fetch(settings[guild.id][0]);
-    oldChannel.delete();
+    let oldChannel = await guild.channels.cache.filter(channel => channel.name.startsWith(channelNames[0])),
+        settings = {
+            type: "voice",
+            permissionOverwrites: [{
+                id: guild.roles.everyone,
+                deny: ['CONNECT']
+            }]
+    };
 
-    let channel = await guild.channels.create(`${channelNames[0]} ${guild.members.cache.filter(member => member.presence.status != "offline").size}`, {
-        type: "voice",
-        parent: settings[guild.id][3],
-        position: 1,
-        permissionOverwrites: [{
-            id: guild.roles.everyone,
-            deny: ['CONNECT']
-            
-        }]
+    oldChannel.each(channel => {
+        if(channel.parentID) settings.parent = channel.parentID;
+
+        settings.position = channel.position;
+
+        channel.delete();
     });
-    settings[guild.id][0] = channel.id;
 
-    fs.writeFileSync(`${__dirname}/settings.json`, JSON.stringify(settings));
+    let channel = await guild.channels.create(`${channelNames[0]} ${guild.members.cache.filter(member => member.presence.status != "offline").size}`, settings);
 }
